Image Gallery on Wearable
Created the ImageCarousel class. Right now is uses 4 pictures stored
within the wear/res/drawable folder. Run Image class on wearable starts
activity, at the moment it is just a Gallery View. Meaning the 4
pictures are displayed on a slider thumbnail size if selected they are
placed in the large view port.