images and base sequence activity
images of clothing plus a start and finish page. I will add a loop to
tranverse a case to swith image buttons source tonight and commit it in
a few hours. I will need Andys help to set the activity to launch but
that will take just a few minute