Time Selection Changes
Time and Date Selection work. However there is a difference between In
App Calendar View vs Online Google Calendar View. I made event for 7:20
start and end 8:20. The In App Calendar Shows it for 6:20 however
checking the Calendar online it shows as the proper 7:20/8:20. This
might be limitations for me because of Emulator. But Works just shows
and hour off between app/calendar.