Simon Says Win Conditions
Added a Win Condition of 10 levels. Simple, Level 1 is 1 button, Level 2
is 2 button pattern and So on. When playing, after Simon's turn is
completed, give it about 3 seconds before clicking the pattern, I think
the user can cause the delays and timers to get out of sync if they are
enter pattern too soon/too fast.